requirements.txt
# Core packages
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0

# NLP & GenAI
transformers>=4.30.0
torch>=2.0.0
nltk>=3.6.0
spacy>=3.5.0
textblob>=0.15.3

# Data collection
tweepy>=4.10.0
praw>=7.6.0
requests>=2.28.0
beautifulsoup4>=4.11.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
streamlit>=1.20.0

# Utilities
python-dotenv>=0.20.0
tqdm>=4.64.0